Move PaymentMethodsV2 component inside the Primary folder.
